requirements.txt
# Core Deep Learning
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=0.24.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress Bars
tqdm>=4.62.0

# Wildlife-specific
wildlife-datasets>=0.1.0

# Interpretability
grad-cam>=1.3.0

# Configuration
pyyaml>=5.4.0
omegaconf>=2.1.0

# Logging and Monitoring
tensorboard>=2.6.0
wandb>=0.12.0

# Utilities
albumentations>=1.0.0
timm>=0.4.12

# Testing (optional, for dev)
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.7b0
flake8>=3.9.0
isort>=5.9.0
mypy>=0.910
pre-commit>=2.15.0
pylint>=2.10.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# API/Web (optional)
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Model Serving (optional)
onnx>=1.10.0
onnxruntime>=1.9.0
